BitNet-archives/BitNet/eval/scripts/requirements.txt
# BitNet Evaluation Framework Requirements

# Core dependencies
numpy>=1.20.0
pyyaml>=6.0

# Dataset loading
datasets>=2.14.0
huggingface_hub>=0.16.0

# Optional: for semantic similarity in stability metrics
# sentence-transformers>=2.2.0

# Optional: for advanced statistics
# scipy>=1.10.0
